Add files via upload
I made a new file with the original code in it.  I did this so that if I deleting some code I knew worked, I could easily go and fix it.  I changed it to incorporate a button that, when pressed, will turn the backlight on on the lcd screen for 5 seconds.